ODFPackage: Use DFStorage, not a filesystem path
An ODFPackage object is our in-memory representation of a .odt, .odp, or
.ods file. Previously, to use this class, you had to first extract the
contents of the .od* file (actually a zip file) in a temporary
directory, so that the ODFPackage load and save routines could then
access the XML files directly from the filesystem.

Now we use DFStorage, which provides a level of abstraction that no
longer requires us to necessarily have the package contents stored in a
temporary directory. A DFStorage object can represent a zip file, a
filesystem directory, or a purely in-memory map of filenames to byte
arrays. The Word filter already works entirely in terms of DFStorage
objects for all its I/O; this commit makes the ODF filter do so too.

ODFPackageCreateNew() is for when you want to create a new ODF package
without any existing data. Although it does not read anything, but just
creates a series of DFDocument objects, it still requires a DFStorage
object to be supplied, as this is retained and used to represent the
location to which the package will eventually be saved. So for example
if you wanted to create a new .odt *file*, you would first call
DFStorageCreateZip(), and then pass the resulting object to
ODFPackageCreateNew().

ODFPackageOpenFrom() is for when you have an existing ODF package
represented by some DFStorage object. It expects to find content.xml,
meta.xml, settings.xml, styles.xml, and META-INF/manifest.xml, parsing
each of these into DFDocument objects. If you want to open an existing
.odt file, you would call DFStorageOpenZip(), and then pass the
resulting object to ODFPackageCreateNew().

ODFPackageSave() serialises the various DFDocument objects to XML and
writes them to the storage object. It then calls DFStorageSave(), to
write them to the actual physical location represented by the storage
object. In the case of DFStorage objects that correspond to zip files,
this causes all the in-memory byte arrays corresponding to the files in
the package to be written to a zip file.