Place cursor properly after deleting an item
When an outline item - either a section, figure, or table - is deleted,
ensure the position is in a valid place with respect to movement (i.e. a
position that the user would normally be allowed to navigate manually).
This is necessary so that we get a valid rect for the cursor, and avoids
an ugly affect whereby the cursor rect would be to the far left of the
screen, and taking up the entire height of the document.