First (non useable) version of Corinthia editor
Added consumers/corinthia as the basis for the new desktop editor based on Qt5.4
CMakeLists.txt is working with Qt, assuming Qt is standard installed.
Source code is copied/modified from a example, and need a lot of massaging.
The editor is stopped for a while, because Qt only runs with 64bit on windows.
Need to fix CMakeLists.txt so the Qt installation + compiler choise can be set
Need to fix 64bit/32bit conflict
In order to test the current code, uncomment #test in the central CMakeList.txt