Added WARNING in consumers/web
To make sure everybody understand the code in this directory and subdirectory,
are preliminary experiments, and will NOT be part of a release.

The code in this directory is used to show proof of concept, and is there to
be shared among the developers who work on the web client editor idea.

Please do not download and use this code, as official Apache code, it is NOT.