Have DFGet() and DFPut() accept a custom id prefix
One of the problems that can occur when updating an existing concrete
document (e.g. a docx file) based on a modified HTML file is that the id
attributes in the latter need to correctly match the sequentially-
assigned node ids in the former. That is, when the put operation goes
through the concrete document working out what to change, it relies on
the mapping from HTML elements to concrete document elements to match
what it was when the original get operation was performed.

It is crucial that this mapping be correct; otherwise, the result can be
corrupted documents due to the wrong elements being moved into the wrong
places, violating the XML document's schema.

A situation that can arise when the conversion library is used for
editing is that a user could open a one document, copy some text to the
clipboard (stored in HTML format, including the id attributes), and then
open a different document and paste that HTML into it. This will result
in the second document containing HTML elements whose id attributes had
meaning in the first document, but not in the latter. When the second
document is later "saved" (that is, its HTML representation is used to
update the originating docx file), these incorrect id attributes can
result in the wrong mapping, causing the problems above.

A solution to this is to use a unique prefix for every conversion, e.g.
a random number. This was previously done, but that capability was taken
out to simplify the API. However, I've received several bug reports from
the latest update to UX Write regarding corrupted documents and/or text
missing on save & reopen, and I suspect the above issue is probably the
cause.

The DFGet() and DFPut() functions now take an optional string to be used
as an id prefix - this can be NULL, in which case the filter will choose
its own (the only current filter this is relevant to is the Word filter,
which chooses the "word" prefix). dfconvert also now allows the user to
optionally specify an id prefix on the command line.