Sort filenames before computing document hash
When computing the hash of all files in a document package, ensure that
we process each file in alphabetical order, to guarantee that the hash
will be the same when this process is done twice for the same document.
There is no guarantee that DFStorageList will return filenames in the
same order each time.

In practice, the current implementation of DFStorageList *does* return
them in the same order, however this is only out of luck. This is due to
the current implementation of DFHashTable, and the way that its contents
are populated when reading a zip file. However, this implementation
could change in in the future, so sorting the results is necessary if we
want to guarantee the ordering.