Cursor placement and selection (mouse only)
Respond to mouse events in Editor, setting the cursor or selection as
appropriate.

To display the cursor, we use a Qt widget which is overlaid on top of
the web view. This is a custom class that simply fills its rectangle
with blue. It his hidden when there is a selection, and visible when
there is no selection.

The logic for rendering the selection is already implemented in
javascript code; it bypasses the selection in WebKit due to bugs on
older verisons of iOS (at the time the code was written; not sure if
they're still present). To render a selection, the JS code creates
<span> elements with a class of "uxwrite-selection" (this should be
"corinthia-selection"), which has a light blue background. These spans
are temporary; when a file is saved (which isn't yet supported in the Qt
version), they will be removed.

From the C++ side, we indicate to the JS code that the cursor or
selection should be adjusted by calling the dragSelectionBegin() and
dragSelectionUpdate() methods on the Selection module of the JS
interface. This causes the appropriate changes to be made to the DOM to
insert <span> elements if a selection is to be displayed, and the
relevant callback methods to be invoked which we use to position and
show/hide the cursor widget.

The EditorJSCallbacks::setCursor() method gets invoked when the JS code
wants us to display a cursor. This callback is triggered after executing
one of the above JS methods, from within the logic in JSInterface.cpp
which checks for any pending callback methods awaiting invocation after
each JS API call.

Note that the JS code considers a "cursor" simply to be an empty
selection - that is, one where the start and end positions are the same.
This is why the handling of both are intertwined.

This commit also amends the sample.html file to explicitly reference the
external stylesheet builtin.css, which is needed for the
uxwrite-selection class. This is done with the following line:

    <link rel="stylesheet" href="builtin.css">

The explicit <link> element is a temporary measure; once we add support
for loading arbitrary files, we will need to dynamically add this to the
DOM so the relevant built-in styles are always available.