Flat: Separate functions for each expression type
Define separate constructor and accessor functions for each type of
expression. This gives us a little more type safety, albeit at runtime
in the form of assertion checks, thus catching any mistakes caused by
trying to access an Expression object in the wrong way for it's type
(e.g. trying to get the string value associated with a sequence).