Flat: Basic recursive-descent parser
Initial implementation of a recursive-descent parser that uses a Grammar
object to product a tree of terms. The parser is very simplistic in
nature; it does a tree walk of the expressions associated with the rules
of the grammar, and does not perform any memoization or other
optimizations.

The resulting Term objects represent the parse tree of the input file.
One Term object is created for each node in the expression tree, and in
the case of character classes and . expressions, one for each character.
The resulting parse trees are thus much larger than one would normally
want, but this is at least a starting point.

Terms and TermLists are not currently freed from memory - the intention
is that these will eventually be handled via garbage collection, as they
will form the set of values that are used for not just parsing but also
any subsequent transformations.