Added functions to work directly with zlib.
Minizip is removed, a.o. due to its problematic license in the main part of corinthia.
New functions have been made, with significant differences:
DFZipOpen, now scans the zip file, and delivers the whole directory back.
DFZipReadFile, reads the whole file in one chunk, and delivers a freshly allocated buffer back
DFZipCreate, creates a new zip file without knowing the content
DFZipWriteFile, adds a whole file to the zip file, and updates the in memory directory.
DFZipClose, releases used memory, and in case of creating a zip file, writes the global directory as well as the end record.

The functions have been tested with LO, AOO, MS-WORD on ODF and OOXML files. The testing showed an error in AOO ODF implementation,
AOO has significant differences between the local header and the global directory, making it very difficult to recover a file with
a corrupted directory. The functions accepts the behaivour of all 6 combinations.