Flat: Label expressions
Add a new expression type, Label, which associated an identifier with
parse tree nodes created from a particular portion of the grammar. This
is useful when you have multiple alternatives in a rule that you want to
distinguish by name, e.g.:

    Term    : $Add(Factor PLUS Term)
            | $Sub(Factor MINUS Term)
            | Factor;
    Factor  : $Mul(Primary TIMES Factor)
            | $Div(Primary SLASH Factor)
            | Primary;

A label expression is written as $SomeLabel(...); it does not affect the
set of possible input strings affected by the parser, but simply
instructs it to create an extra node in the parse tree with the label
"SomeLabel". In the above example, taken from arithmetic.flat, each type
of expression is labeled with the appropriate constructor.

The parse trees we've created so far have all contained lots of nodes,
particularly for intermediate expressions like rule references, choices,
sequences, optional types, and lists. With explicit labels/constructors
as above, we are now part way to being able to construct simpler parse
trees which only store the semantically-important information in the
abstract syntax tree.