included MOC compiler for Qt implementation
Changed the Layer namings, to avoid the PITA of having all layers start with Layer<foo>.
Added moc generated moc file to the Qt implementation (this is to avoid everybody needing to run moc)
Ready to all graphic in the Qt implementation.
The Web implementation has been verified on paper, but will not be implemented until Qt works.