Comments from IPMC on release 0.1
DocFormats-Info.plist and DocFormats-Prefix.pch removed, because they do seem unused
(and misses ALv2 header).

LINK_DEPENDENCIES.txt got a comment, so that is moved into README.md instead.

We are recommed to have the actual build instructions in the source zip, since they might
change with time, and we only have current ones in wiki.